fix: 패키지 merge, bulk delete 추가
