--- src/main/java/com/polzzak/domain/mission/dto/MissionCompleteCreateRequest.java
@@ -1,10 +1,0 @@
-package com.polzzak.domain.mission.dto;
-
-import jakarta.annotation.Nonnull;
-
-public record MissionCompleteCreateRequest(
-	@Nonnull Long stampBoardId,
-	@Nonnull Long missionId,
-	@Nonnull Long guardianId
-) {
-}
