--- src/main/java/com/polzzak/domain/mission/dto/MissionDto.java
@@ -1,10 +1,0 @@
-package com.polzzak.domain.mission.dto;
-
-import com.polzzak.domain.mission.entity.Mission;
-
-public record MissionDto(Long id, String content) {
-
-	public static MissionDto from(Mission mission) {
-		return new MissionDto(mission.getId(), mission.getContent());
-	}
-}
