--- src/main/java/com/polzzak/domain/mission/repository/MissionCompleteRepository.java
@@ -1,8 +1,0 @@
-package com.polzzak.domain.mission.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.polzzak.domain.mission.entity.MissionComplete;
-
-public interface MissionCompleteRepository extends JpaRepository<MissionComplete, Long> {
-}
