--- src/main/java/com/polzzak/domain/stamp/dto/StampBoardGroup.java
@@ -1,10 +1,0 @@
-package com.polzzak.domain.stamp.dto;
-
-public enum StampBoardGroup {
-
-	IN_PROGRESS, ENDED;
-
-	public static StampBoardGroup getStampBoardGroupByStr(String stampBoardGroupAsStr) {
-		return valueOf(stampBoardGroupAsStr.toUpperCase());
-	}
-}
