--- src/main/java/com/polzzak/domain/stamp/dto/StampCreateRequest.java
@@ -1,7 +1,0 @@
-package com.polzzak.domain.stamp.dto;
-
-import jakarta.annotation.Nonnull;
-import jakarta.validation.constraints.Min;
-
-public record StampCreateRequest(@Min(1) int count, @Nonnull Long missionId, @Nonnull Integer stampDesignId) {
-}
