--- src/main/java/com/polzzak/domain/stamp/entity/Stamp.java
@@ -1,57 +1,0 @@
-package com.polzzak.domain.stamp.entity;
-
-import java.time.LocalDateTime;
-
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import com.polzzak.domain.mission.entity.Mission;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToOne;
-import jakarta.persistence.Table;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Entity
-@NoArgsConstructor
-@EntityListeners(AuditingEntityListener.class)
-@Table(name = "stamp")
-public class Stamp {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "stamp_board_id")
-	private StampBoard stampBoard;
-
-	@Column(nullable = false)
-	private int stampDesignId;
-
-	@OneToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "mission_id")
-	private Mission mission;
-
-	@CreatedDate
-	@Column(nullable = false, updatable = false)
-	private LocalDateTime createdDate;
-
-	@Builder(builderMethodName = "createMission")
-	public Stamp(StampBoard stampBoard, Mission mission, int stampDesignId) {
-		this.stampBoard = stampBoard;
-		this.mission = mission;
-		this.stampDesignId = stampDesignId;
-	}
-}
