+++ src/main/java/com/polzzak/domain/stampboard/dto/MissionCompleteCreateRequest.java
@@ -1,0 +1,10 @@
+package com.polzzak.domain.stampboard.dto;
+
+import jakarta.annotation.Nonnull;
+
+public record MissionCompleteCreateRequest(
+	@Nonnull Long stampBoardId,
+	@Nonnull Long missionId,
+	@Nonnull Long guardianId
+) {
+}
