+++ src/main/java/com/polzzak/domain/stampboard/dto/MissionCompleteDto.java
@@ -1,0 +1,17 @@
+package com.polzzak.domain.stampboard.dto;
+
+import java.time.LocalDateTime;
+
+import com.polzzak.domain.stampboard.entity.MissionComplete;
+
+public record MissionCompleteDto(
+	long id,
+	String missionContent,
+	LocalDateTime createdDate
+) {
+
+	public static MissionCompleteDto from(MissionComplete missionComplete) {
+		return new MissionCompleteDto(missionComplete.getId(), missionComplete.getMission().getContent(),
+			missionComplete.getCreatedDate());
+	}
+}
