+++ src/main/java/com/polzzak/domain/stampboard/dto/StampBoardGroup.java
@@ -1,0 +1,10 @@
+package com.polzzak.domain.stampboard.dto;
+
+public enum StampBoardGroup {
+
+	IN_PROGRESS, ENDED;
+
+	public static StampBoardGroup getStampBoardGroupByStr(String stampBoardGroupAsStr) {
+		return valueOf(stampBoardGroupAsStr.toUpperCase());
+	}
+}
